requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
comm==0.1.3
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
discord.py==2.3.1
dnspython==2.3.0
executing==1.2.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.273
langsmith==0.0.26
loguru==0.7.0
lxml==4.9.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mmh3==4.0.0
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.2
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.11
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
ruff==0.0.278
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
tenacity==8.2.2
tiktoken==0.4.0
tokenize-rt==5.1.0
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
wget==3.2
yarl==1.9.2
